feat: rename mod curl to async_curl
This is to avoid conflict if we will expose curl
public structs and enums
